add: added logout functionality and redirects for non logged in users

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -25,6 +27,10 @@ from .forms import RoomForm
 
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == "POST":
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -45,6 +51,12 @@ def loginPage(request):
     template = 'base/login_register.html'
     context = {}
     return render(request, template, context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
 
 def home(request):
 
@@ -78,6 +90,7 @@ def room(request, pk):
     return render(request, template, context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -93,9 +106,14 @@ def createRoom(request):
     return render(request, template, context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here.')
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -108,8 +126,14 @@ def updateRoom(request, pk):
 
     return render(request, template, context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here.')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
